rocket_launch module

+++ src/modules/rocket_launch/actions.cpp
@@ -26,7 +26,8 @@ ScheduleLaunchAction::validate(const flecs::world &world) const {
   if (!launchpad.is_valid()) {
     return ValidationResult::Fail("No launchpad selected");
   }
-  unsigned int launchPrepDays = launchpad.get<Launchpad>().prep_days.value();
+  uint32_t launchPrepDays =
+      static_cast<uint32_t>(launchpad.get<Launchpad>().prep_days.value());
   bool clash = false;
   launchpad.each<LaunchingFrom>([&](flecs::entity p) {
     auto launch = p.get<LaunchPlan>();
